requirements.txt
torch
torchvision
Pillow
psutil
pynvml
pandas
numpy
scikit-learn
matplotlib